requirements.txt
PyYAML>=6.0
tqdm>=4.66
numpy>=1.26
matplotlib>=3.8
